Add first pass at a module on graphing with matplotlib.
The module has 3 major exercises:

1. test.py: generate a basic dot plot.

2. world_population.py: building on the pyplot functions covered in
test.py, plot the world population over time.

3. constitution.py: create a bar chart for the letter frequencies in
the US Constitution.

The module comes with a suggested workflow for instructors, and
suggested exercises. One of the suggested exercises is to plot the
data in life_expectancies_usa.txt. An example implementation is in
life_expectancies_usa.DONE.py.